feat: remove start page (#376)

+++ src/jsx/parts/menu-bar.jsx
@@ -1,16 +1,22 @@
 import { h, Component } from 'preact'
 
 import Button from '../parts/button.jsx'
+import Loading from '../parts/loading.jsx'
 
 
 class MenuBar extends Component {
   constructor(props) {
     super(props)
     this.state = {
+      url: '',
       isDeleting: false,
+      isCreating: false,
     }
+    this.creating = this.creating.bind(this)
     this.deleting = this.deleting.bind(this)
     this.cancel = this.cancel.bind(this)
+    this.create = this.create.bind(this)
+    this.setUrl = this.setUrl.bind(this)
   }
 
   componentWillReceiveProps({ id }) {
@@ -19,9 +25,11 @@ class MenuBar extends Component {
     }
   }
 
-  shouldComponentUpdate({ id, isSelectorOpen }, { isDeleting }) {
+  shouldComponentUpdate({ id, loading, isSelectorOpen }, { isCreating, isDeleting }) {
     return !(this.props.id === id
+        && this.props.loading === loading
         && this.props.isSelectorOpen === isSelectorOpen
+        && this.state.isCreating === isCreating
         && this.state.isDeleting === isDeleting)
   }
 
@@ -29,26 +37,51 @@ class MenuBar extends Component {
     this.setState({ isDeleting: true })
   }
 
+  creating() {
+    this.setState({ isCreating: true })
+  }
+
   cancel() {
-    this.setState({ isDeleting: false })
+    this.setState({
+      url: '',
+      isDeleting: false,
+      isCreating: false
+    })
+  }
+
+  setUrl(event) {
+    this.setState({ url: event.target.value })
+  }
+
+  create() {
+    this.props.setLoading()
+    this.props.getFile(this.state.url)
   }
 
   render({
     id,
-    cancel,
+    loading,
     deleteOne,
-    deleting,
-    fileUrl,
     searchCodeAndComment,
     searchComment,
     toggleSelector,
-    isSelectorOpen
+    isSelectorOpen,
+    networkError,
+    urlError
   }, {
+    isCreating,
     isDeleting,
   }) {
-    if (isDeleting) {
+    if (loading) {
+      return(
+        <div className="cc-menu-bar loading">
+          <Loading />
+        </div>
+      )
+    }
+    else if (isDeleting) {
       return (
-        <div key="menu-bar" className="cc-menu-bar deleting">
+        <div className="cc-menu-bar deleting">
           <div className="message">This code and comment is removed.</div>
           <Button onClick={ deleteOne }>OK</Button>
           { ' ' }
@@ -56,11 +89,23 @@ class MenuBar extends Component {
         </div>
       )
     }
+    else if (isCreating) {
+      return (
+        <div className="cc-menu-bar creating">
+          <p>Input the file url in Github.</p>
+          <input type="text" className="url" onChange={ this.setUrl }/>
+          <Button onClick={ this.create }>Create</Button>
+          <Button onClick={ this.cancel }>Cancel</Button>
+          { networkError && <div>The file data is not got.</div> }
+          { urlError && <div>Url is invalid.</div> }
+        </div>
+      )
+    }
     else {
       return (
-        <div key="menu-bar" className="cc-menu-bar">
+        <div className="cc-menu-bar">
           <span className="label" onClick={ toggleSelector }>{ isSelectorOpen ? 'Close' : 'Open' }</span>
-          <span className="label" onClick={ fileUrl }>New</span>
+          <span className="label" onClick={ this.creating }>New</span>
           <span className="label" onClick={ searchCodeAndComment }>List</span>
           <span className="label" onClick={ searchComment }>Comments</span>
           { id && <span className="label" onClick={ this.deleting }>Delete</span> }
